+++ client/dashboard/sites/site/switcher.tsx
@@ -8,11 +8,18 @@ import { sitesQuery } from '../../app/queries/sites';
 import RouterLinkMenuItem from '../../components/router-link-menu-item';
 import AddNewSite from '../add-new-site';
 import SiteIcon from '../site-icon';
+import type { Site } from '../../data/types';
 import type { View } from '@automattic/dataviews';
 
 import './switcher.scss';
 
-const fields = [ { id: 'name', enableGlobalSearch: true } ];
+const fields = [
+	{
+		id: 'name',
+		getValue: ( { item }: { item: Site } ) => item.name || new URL( item.URL ).hostname,
+		enableGlobalSearch: true,
+	},
+];
 
 const DEFAULT_VIEW: View = {
 	type: 'list',
@@ -51,7 +58,7 @@ export default function Switcher( { onClose }: { onClose: () => void } ) {
 							<span
 								style={ { overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap' } }
 							>
-								{ site.name }
+								{ fields[ 0 ].getValue( { item: site } ) }
 							</span>
 						</div>
 					</RouterLinkMenuItem>