+++ client/landing/stepper/declarative-flow/internals/steps-repository/__user/index.tsx
@@ -43,7 +43,10 @@ const UserStepComponent: StepType = function UserStep( {
 			wpcom.loadToken( wpAccountCreateResponse.bearer_token );
 			reloadProxy();
 			requestAllBlogsAccess();
-			dispatch( fetchCurrentUser() as unknown as AnyAction );
+			// Allow retries of fetching new users after creation. New user sign-ups go to one DC
+			// but follow-up API calls go to the closest DC, which may be different and might not
+			// have replicated the user data yet.
+			dispatch( fetchCurrentUser( { retry: true } ) as unknown as AnyAction );
 		}
 		if ( ! isLoggedIn ) {
 			dispatch( fetchCurrentUser() as unknown as AnyAction );