+++ packages/dataviews/src/components/dataviews-filters/filter.tsx
@@ -47,6 +47,7 @@ import {
 	OPERATOR_AFTER,
 	OPERATOR_BEFORE_INC,
 	OPERATOR_AFTER_INC,
+	OPERATOR_BETWEEN,
 } from '../../constants';
 import type {
 	Filter,
@@ -299,6 +300,23 @@ const FilterText = ( {
 		);
 	}
 
+	if ( filterInView?.operator === OPERATOR_BETWEEN ) {
+		const { label } = activeElements[ 0 ];
+
+		return createInterpolateElement(
+			sprintf(
+				/* translators: 1: Filter name. 2: Min value. 3: Max value. e.g.: "Item count between (inc): 10-180". */
+				__(
+					'<Name>%1$s between (inc): </Name><Value>%2$s-%3$s</Value>'
+				),
+				filter.name,
+				label[ 0 ],
+				label[ 1 ]
+			),
+			filterTextWrappers
+		);
+	}
+
 	return sprintf(
 		/* translators: 1: Filter name e.g.: "Unknown status for Author". */
 		__( 'Unknown status for %1$s' ),
@@ -337,6 +355,7 @@ function OperatorSelector( {
 					options={ operatorOptions }
 					onChange={ ( newValue ) => {
 						const operator = newValue as Operator;
+						const currentOperator = currentFilter?.operator;
 						const newFilters = currentFilter
 							? [
 									...( view.filters ?? [] ).map(
@@ -344,8 +363,18 @@ function OperatorSelector( {
 											if (
 												_filter.field === filter.field
 											) {
+												// Reset the value only when switching between "between" and any other operator to avoid invalid values.
+												const isSwitchingBetween =
+													currentOperator ===
+														OPERATOR_BETWEEN ||
+													operator ===
+														OPERATOR_BETWEEN;
+
 												return {
 													..._filter,
+													value: isSwitchingBetween
+														? undefined
+														: _filter.value,
 													operator,
 												};
 											}