+++ packages/ui/src/calendar/date-range-calendar/test/index.tsx
@@ -1,7 +1,3 @@
-/**
- * @jest-environment jsdom
- */
-
 import { render, screen, within, renderHook } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import {
@@ -74,14 +70,14 @@ function setupUserEvent() {
 }
 
 describe( 'DateRangeCalendar', () => {
-	let today;
-	let tomorrow;
-	let yesterday;
-	let currentMonth;
-	let nextMonth;
-	let nextNextMonth;
-	let prevMonth;
-	let prevPrevMonth;
+	let today: Date;
+	let tomorrow: Date;
+	let yesterday: Date;
+	let currentMonth: Date;
+	let nextMonth: Date;
+	let nextNextMonth: Date;
+	let prevMonth: Date;
+	let prevPrevMonth: Date;
 
 	beforeAll( () => {
 		jest.useFakeTimers();