+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+## 1.0.0
+
 ### Enhancements
 
-- `Badge`: Add component (forked from `@wordpress/components`).
+- `Badge`: Add component (forked from `@wordpress/components`) ([#104187](https://github.com/Automattic/wp-calypso/pull/104187)).
+- `DateCalendar` and `DateRangeCalendar`: Add components ([#104224](https://github.com/Automattic/wp-calypso/pull/104224)).